# doc/api.yaml
openapi: 3.0.0
paths:
  /user:
    get:
      description: Gets all users
      operationId: UserController_findAll
      parameters: []
      responses:
        '200':
          description: Returns all users records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
      summary: Get all users
      tags:
        - User
    post:
      description: Creates a new user
      operationId: UserController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: Returns if user has been created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Returns if request body does not contain required fields
      summary: Create user
      tags:
        - User
  /user/{id}:
    get:
      description: Get single user by id
      operationId: UserController_findOne
      parameters:
        - name: id
          required: true
          in: path
          description: User ID
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Returns if user record exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Returns if user ID is not valid UUID
        '404':
          description: Returns if user record does't exist
      summary: Get single user by id
      tags:
        - User
    put:
      description: Updates a user's password by ID
      operationId: UserController_updatePassword
      parameters:
        - name: id
          required: true
          in: path
          description: User ID
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasswordDto'
      responses:
        '200':
          description: Returns if user has been updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Returns if user ID is not valid UUID
        '403':
          description: Returns if old password is wrong
        '404':
          description: Returns if user does not exist
      summary: Update a user's password
      tags:
        - User
    delete:
      description: Deletes user by ID
      operationId: UserController_remove
      parameters:
        - name: id
          required: true
          in: path
          description: User ID
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: Returns if user has been removed
        '400':
          description: Returns if user ID is not valid UUID
        '404':
          description: Returns if user does not exist
      summary: Delete user
      tags:
        - User
...
info:
  title: Home Library Service
  description: Home music library service
  version: '1.0'
  contact: {}
...
servers: []
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        login:
          type: string
          example: TestUser
        version:
          type: number
          example: 1
        createdAt:
          type: number
          description: Timestamp of creation
          example: 1655000000
        updatedAt:
          type: number
          description: Timestamp of last update
          example: 1655000000
      required:
        - id
        - login
        - version
        - createdAt
        - updatedAt
    CreateUserDto:
      type: object
      properties:
        login:
          type: string
          description: The user's login
        password:
          type: string
          description: The user's password
      required:
        - login
        - password
    UpdatePasswordDto:
      type: object
      properties:
        oldPassword:
          type: string
          description: The user's old password
        newPassword:
          type: string
          description: The user's new password
      required:
        - oldPassword
        - newPassword
...